requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.11
groq==0.19.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
